Add `pubkey` macro
This macro converts a base58 string to a `Pubkey` instance. It allows to
assign the resulting `Pubkey` as `const` and does the conversion during
compilation (so there is no performance penalty on runtime).

The motivation behind introducing that macro is the fact that
`crate::ID` generated by Anchor's `declare_id` macro is `static`, so it
cannot be assigned to constants.