Fix `light_verifier_accounts` macro
Before this change, using this macro in real crates was resulting in
`not a non-macro attribute errors`. That's because `#[derive(Accounts)]`
macro in anchor is a proc-derive macro, which uses `accounts` as an
attribute.

However, we can't implement our `light_verifier_accounts` macro as a
proc-derive macro, because we are not able to append fields to the
original struct definition through it. To be able to properly modify the
body of the struct.